+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/MessageListState.kt
@@ -17,6 +17,7 @@
 package io.getstream.chat.android.ui.utils.extensions
 
 import io.getstream.chat.android.ui.common.state.messages.list.MessageListState
+import io.getstream.chat.android.ui.common.state.messages.list.NewMessageState
 import io.getstream.chat.android.ui.feature.messages.list.adapter.MessageListItem
 import io.getstream.chat.android.ui.model.MessageListItemWrapper
 
@@ -24,10 +25,14 @@ import io.getstream.chat.android.ui.model.MessageListItemWrapper
  * Converts the common [MessageListState] to ui-components [MessageListItemWrapper].
  *
  * @param isInThread Whether the message list is currently in thread mode or not.
+ * @param prevNewMessageState The previous [NewMessageState] to compare with the current one.
  *
  * @return [MessageListItemWrapper] derived from [MessageListState].
  */
-public fun MessageListState.toMessageListItemWrapper(isInThread: Boolean): MessageListItemWrapper {
+public fun MessageListState.toMessageListItemWrapper(
+    isInThread: Boolean,
+    prevNewMessageState: NewMessageState? = null,
+): MessageListItemWrapper {
     var messagesList: List<MessageListItem> = messageItems.map { it.toUiMessageListItem() }
 
     if (isLoadingOlderMessages) messagesList = messagesList + listOf(MessageListItem.LoadingMoreIndicatorItem)
@@ -35,7 +40,8 @@ public fun MessageListState.toMessageListItemWrapper(isInThread: Boolean): Messa
 
     return MessageListItemWrapper(
         items = messagesList,
-        hasNewMessages = newMessageState != null,
+        hasNewMessages = newMessageState != prevNewMessageState,
+        newMessageState = newMessageState,
         isTyping = messagesList.firstOrNull { it is MessageListItem.TypingItem } != null,
         areNewestMessagesLoaded = endOfNewMessagesReached,
         isThread = isInThread,